# Refactor Router#find_routes to be lazier